Removed the clutter
    - Removed zproto_example_go.xml, zproto_example_goczmq.xml and
      zproto_example_go.xml
    - Moved goczmq to go/
    - Changed zproto_lib_go.gsl and zproto_lib_php.gsl to respect switches
    - Modified Makefile.am to use switches to generate go and php codecs
