+++ src/Makefile.am
@@ -27,6 +27,8 @@ zproto_DATA = \
 # Produce generated models; do this manually in src directory
 code:
 	gsl -q -script:zproto_codec_c zproto_example.xml
-	gsl -q -script:zproto_codec_go zproto_example.xml
+	gsl -q -script:zproto_codec_go -package_dir:"go/zmq4" -package:"github.com/pebbe/zmq4" zproto_example.xml
+	gsl -q -script:zproto_codec_go -package_dir:"go/goczmq" -package:"github.com/zeromq/goczmq" zproto_example.xml
+	gsl -q -script:zproto_codec_php -package_dir:php -test_dir:php -namespace:Zproto zproto_example.xml
 	gsl -q -script:zproto_codec_java zproto_example.xml
 	gsl -q -script:zproto_codec_clj zproto_example.xml
