+++ src/go/goczmq/zproto/example/binary.go
