+++ src/go/goczmq/zproto/example/binary_test.go
