+++ src/go/goczmq/zproto/example/log.go
