+++ src/go/goczmq/zproto/example/log_test.go
