+++ src/go/goczmq/zproto/example/structures.go
