+++ src/go/goczmq/zproto/example/structures_test.go
