+++ src/go/goczmq/zproto/example/types.go
