+++ src/go/zmq4/zproto/example/binary.go
