+++ src/go/zmq4/zproto/example/binary_test.go
