+++ src/go/zmq4/zproto/example/log.go
