+++ src/go/zmq4/zproto/example/log_test.go
