+++ src/go/zmq4/zproto/example/structures.go
