+++ src/go/zmq4/zproto/example/structures_test.go
