+++ src/php/Example/Binary.php
@@ -6,7 +6,7 @@
  * DO NOT MAKE ANY CHANGES YOU WISH TO KEEP.
  *
  * The correct places for commits are:
- *  - The XML model used for this code generation: zproto_example_php.xml
+ *  - The XML model used for this code generation: zproto_example.xml
  *  - The code generation script that built this file: zproto_codec_php
  *
  *  Copyright (C) 2014 the Authors
